Added promise await in DIC
Services should be autoloaded solved before the first request is served.
This means that we could take in consideration to await all Promises to
be solved before start handling requests. This is why we can select a
DIC service as `await` by using a simple tag, and the container will
await the promise to be fulfilled or rejected.

```php
My\Future\Service:
    tags:
        - {"name": "async"}
```